[#9710] rename to criteria and audit
