+++ sormas-api/src/main/java/de/symeda/sormas/api/task/TaskContextIndexCriteria.java
@@ -1,0 +1,29 @@
+package de.symeda.sormas.api.task;
+
+import de.symeda.sormas.api.audit.AuditInclude;
+import de.symeda.sormas.api.audit.AuditedClass;
+
+import java.io.Serializable;
+
+@AuditedClass
+public class TaskContextIndexCriteria implements Serializable, Cloneable {
+
+	private TaskContext taskContext;
+	@AuditInclude
+	private String uuid;
+
+	public TaskContextIndexCriteria() {
+	}
+
+	public TaskContextIndexCriteria(TaskContext taskContext) {
+		this.taskContext = taskContext;
+	}
+
+	public TaskContext getTaskContext() {
+		return taskContext;
+	}
+
+	public String getUuid() {
+		return uuid;
+	}
+}
